packages/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
sqlalchemy
psycopg2-binary
email-validator
python-multipart